+++ application.go
@@ -6,8 +6,8 @@ import (
 
 	"github.com/dogmatiq/configkit/internal/typename/goreflect"
 	"github.com/dogmatiq/configkit/internal/validation"
-	"github.com/dogmatiq/configkit/message"
 	"github.com/dogmatiq/dogma"
+	"github.com/dogmatiq/enginekit/message"
 )
 
 // Application is an interface that represents the configuration of a Dogma