+++ message/doc.go
@@ -1,5 +1,0 @@
-// Package message provides utilities for representing information about Dogma
-// messages and their use within an application.
-//
-// Deprecated: Use [github.com/dogmatiq/enginekit/message] instead.
-package message