+++ static/entity.go
@@ -4,7 +4,7 @@ import (
 	"context"
 
 	"github.com/dogmatiq/configkit"
-	"github.com/dogmatiq/configkit/message"
+	"github.com/dogmatiq/enginekit/message"
 )
 
 // application is an implementation of [configkit.Application].