+++ visualization/dot/style.go
@@ -2,7 +2,7 @@ package dot
 
 import (
 	"github.com/dogmatiq/configkit"
-	"github.com/dogmatiq/configkit/message"
+	"github.com/dogmatiq/enginekit/message"
 	"github.com/emicklei/dot"
 )